Moved `database.py` into a folder
