+++ website/config.py
@@ -1,7 +1,8 @@
 """
 This file contains all the data needed for the game
 """
-from .database import Player
+
+from .database.database import Player
 import copy
 import math
 from flask import current_app
