+++ website/utils.py
@@ -16,7 +16,7 @@ import json
 
 
 from website.api import ws
-from .database import (
+from .database.database import (
     Player,
     Network,
     Resource_on_sale,
