+++ website/views.py
@@ -6,7 +6,7 @@ from flask import Blueprint, render_template, request, flash
 from flask import g, current_app
 from flask_login import login_required, current_user
 from . import db
-from .database import Chat, Player, Resource_on_sale
+from .database.database import Chat, Player, Resource_on_sale
 from .utils import check_existing_chats
 
 views = Blueprint("views", __name__)
